requirements.txt
flask
gunicorn
deta
python-dotenv